Add timeout command, user context & msg context
This commit adds custom timeouts, through a command, a user context
menu option, AND a message context menu option.

This is very useful for custom durations, since Discord's UI only
allows a limited amount of them. Additionally, this DMs the user the
reason.

This commit also switches the bot to ESM (from CommonJS).